refactor : main page - newBook, themeRecommendation component
신규도서 및 테마추천 컴포넌트 세분화
메인페이지 컴포넌트 폴더 구조 리팩토링
메인페이지에서 생성한 타 페이지 공유 컴포넌트 폴더 구조 리팩토링

Related to : #18
