--- src/components/common/ex.txt
@@ -1,1 +1,0 @@
-common: 2회 이상 재사용되는 공용 컴포넌트 파일
