--- src/components/ex.txt
@@ -1,1 +1,0 @@
-각 페이지에서 생성된 컴포넌트 파일
