+++ src/styles/main/themeRecommendation/themeItem.module.css
@@ -1,0 +1,37 @@
+.themeItem {
+	width: 400px;
+	height: 200px;
+	border-radius: 10px;
+
+	position: relative;
+	overflow: hidden;
+
+	background-color: #993366;
+}
+.themeItem:hover > .themeImage {
+	transform: scale(1.05);
+}
+.themeText {
+	width: 225px;
+	height: 200px;
+	gap: 10px;
+
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	flex-direction: column;
+
+	color: #ffffff;
+}
+.themeImage {
+	width: 150px;
+	height: 200px;
+	top: 50px;
+	left: 225px;
+	border-radius: 10px;
+
+	position: absolute;
+	overflow: hidden;
+
+	transition: transform 0.3s ease-in-out;
+}
