+++ src/styles/main/themeRecommendation/themeRecommendation.module.css
@@ -1,0 +1,18 @@
+.themeRecommendationWrapper {
+	width: inherit;
+	height: 510px;
+
+	display: flex;
+	flex-direction: column;
+}
+
+.themeItemWrapper {
+	width: 1300px;
+	height: 450px;
+	gap: 47px;
+
+	display: flex;
+	flex-wrap: wrap;
+	justify-content: space-between;
+	overflow: hidden;
+}
